feat: 역 삭제시 참조되는 역은 삭제 불가능 구현

+++ src/main/java/wooteco/subway/infrastructure/station/StationRepositoryImpl.java
@@ -30,8 +30,13 @@ public class StationRepositoryImpl implements StationRepository {
         return stationDao.findAll();
     }
 
+    //is it possible??
     @Override
     public int delete(Long id) {
+        if(stationDao.isReferenced(id)) {
+            throw new IllegalArgumentException("해당 역을 참조하는 노선이 있어 삭제가 불가능합니다.");
+        }
+
         return stationDao.delete(id);
     }
 }
